moved AnecdoteList for listing and voting to a component
